requirements.txt
aniso8601==8.0.0
click==7.1.2
dateparser==0.7.6
dnspython==1.16.0
eventlet==0.28.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
greenlet==0.4.17
humanize==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
maya==0.6.1
monotonic==1.5
pendulum==2.1.2
Pillow==8.0.0
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
pytzdata==2020.1
regex==2020.10.15
six==1.15.0
snaptime==0.2.4
tzlocal==2.1
Werkzeug==1.0.1
